Move files to proto/primitive and use tbots_protobuf funcs
