+++ src/software/primitive/BUILD
@@ -153,28 +153,6 @@ cc_library(
 )
 
 cc_library(
-    name = "primitive_msg_factory",
-    srcs = ["primitive_msg_factory.cpp"],
-    hdrs = ["primitive_msg_factory.h"],
-    deps = [
-        ":primitive",
-        "//shared/proto:tbots_cc_proto",
-        "//software/geom:angle",
-        "//software/geom:angular_velocity",
-        "//software/geom:point",
-    ],
-)
-
-cc_test(
-    name = "primitive_msg_factory_test",
-    srcs = ["primitive_msg_factory_test.cpp"],
-    deps = [
-        ":primitive_msg_factory",
-        "@gtest//:gtest_main",
-    ],
-)
-
-cc_library(
     name = "primitive_visitor",
     hdrs = ["primitive_visitor.h"],
 )
