+++ src/software/proto/primitive/BUILD
@@ -1,0 +1,26 @@
+load("@rules_proto//proto:defs.bzl", "proto_library")
+
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "primitive_msg_factory",
+    srcs = ["primitive_msg_factory.cpp"],
+    hdrs = ["primitive_msg_factory.h"],
+    deps = [
+        "//shared/proto:tbots_cc_proto",
+        "//software/geom:angle",
+        "//software/geom:angular_velocity",
+        "//software/geom:point",
+        "//software/primitive",
+        "//software/proto/message_translation:tbots_protobuf",
+    ],
+)
+
+cc_test(
+    name = "primitive_msg_factory_test",
+    srcs = ["primitive_msg_factory_test.cpp"],
+    deps = [
+        ":primitive_msg_factory",
+        "@gtest//:gtest_main",
+    ],
+)
