+++ src/software/proto/primitive/primitive_msg_factory.cpp
@@ -1,4 +1,6 @@
-#include "software/primitive/primitive_msg_factory.h"
+#include "software/proto/primitive/primitive_msg_factory.h"
+
+#include "software/proto/message_translation/tbots_protobuf.h"
 
 std::unique_ptr<ChipPrimitiveMsg> createChipPrimitiveMsg(const Point &chip_origin,
                                                          const Angle &chip_direction,
@@ -6,13 +8,10 @@ std::unique_ptr<ChipPrimitiveMsg> createChipPrimitiveMsg(const Point &chip_origi
 {
     auto chip_primitive_msg = std::make_unique<ChipPrimitiveMsg>();
 
-    auto chip_origin_msg = std::make_unique<PointMsg>();
-    chip_origin_msg->set_x_meters(static_cast<float>(chip_origin.x()));
-    chip_origin_msg->set_y_meters(static_cast<float>(chip_origin.y()));
+    auto chip_origin_msg = createPointMsg(Point(chip_origin.x(), chip_origin.y()));
     *(chip_primitive_msg->mutable_chip_origin()) = *chip_origin_msg;
 
-    auto chip_direction_msg = std::make_unique<AngleMsg>();
-    chip_direction_msg->set_radians(static_cast<float>(chip_direction.toRadians()));
+    auto chip_direction_msg                         = createAngleMsg(chip_direction);
     *(chip_primitive_msg->mutable_chip_direction()) = *chip_direction_msg;
 
     chip_primitive_msg->set_chip_distance_meters(chip_distance_meters);
@@ -26,13 +25,10 @@ std::unique_ptr<KickPrimitiveMsg> createKickPrimitiveMsg(
 {
     auto kick_primitive_msg = std::make_unique<KickPrimitiveMsg>();
 
-    auto kick_origin_msg = std::make_unique<PointMsg>();
-    kick_origin_msg->set_x_meters(static_cast<float>(kick_origin.x()));
-    kick_origin_msg->set_y_meters(static_cast<float>(kick_origin.y()));
+    auto kick_origin_msg = createPointMsg(Point(kick_origin.x(), kick_origin.y()));
     *(kick_primitive_msg->mutable_kick_origin()) = *kick_origin_msg;
 
-    auto kick_direction_msg = std::make_unique<AngleMsg>();
-    kick_direction_msg->set_radians(static_cast<float>(kick_direction.toRadians()));
+    auto kick_direction_msg                         = createAngleMsg(kick_direction);
     *(kick_primitive_msg->mutable_kick_direction()) = *kick_direction_msg;
 
     kick_primitive_msg->set_kick_speed_meters_per_second(kick_speed_meters_per_second);
@@ -47,16 +43,13 @@ std::unique_ptr<MovePrimitiveMsg> createMovePrimitiveMsg(
     auto move_primitive_msg = std::make_unique<MovePrimitiveMsg>();
 
     auto position_params_msg = std::make_unique<MovePositionParams>();
-    auto dest_msg            = std::make_unique<PointMsg>();
-    dest_msg->set_x_meters(static_cast<float>(dest.x()));
-    dest_msg->set_y_meters(static_cast<float>(dest.y()));
+    auto dest_msg            = createPointMsg(Point(dest.x(), dest.y()));
     *(position_params_msg->mutable_destination()) = *dest_msg;
     position_params_msg->set_final_speed_meters_per_second(final_speed_meters_per_second);
     position_params_msg->set_slow(slow);
     *(move_primitive_msg->mutable_position_params()) = *position_params_msg;
 
-    auto final_angle_msg = std::make_unique<AngleMsg>();
-    final_angle_msg->set_radians(static_cast<float>(final_angle.toRadians()));
+    auto final_angle_msg                         = createAngleMsg(final_angle);
     *(move_primitive_msg->mutable_final_angle()) = *final_angle_msg;
 
     move_primitive_msg->set_dribbler_speed_rpm(dribbler_speed_rpm);
@@ -71,17 +64,13 @@ std::unique_ptr<SpinningMovePrimitiveMsg> createSpinningMovePrimitiveMsg(
     auto spinning_move_primitive_msg = std::make_unique<SpinningMovePrimitiveMsg>();
 
     auto position_params_msg = std::make_unique<MovePositionParams>();
-    auto dest_msg            = std::make_unique<PointMsg>();
-    dest_msg->set_x_meters(static_cast<float>(dest.x()));
-    dest_msg->set_y_meters(static_cast<float>(dest.y()));
+    auto dest_msg            = createPointMsg(Point(dest.x(), dest.y()));
     *(position_params_msg->mutable_destination()) = *dest_msg;
     position_params_msg->set_final_speed_meters_per_second(final_speed_meters_per_second);
     position_params_msg->set_slow(slow);
     *(spinning_move_primitive_msg->mutable_position_params()) = *position_params_msg;
 
-    auto angular_velocity_msg = std::make_unique<AngularVelocityMsg>();
-    angular_velocity_msg->set_radians_per_second(
-        static_cast<float>(angular_velocity.toRadians()));
+    auto angular_velocity_msg = createAngularVelocityMsg(angular_velocity);
     *(spinning_move_primitive_msg->mutable_angular_velocity()) = *angular_velocity_msg;
 
     spinning_move_primitive_msg->set_dribbler_speed_rpm(dribbler_speed_rpm);
@@ -96,16 +85,13 @@ std::unique_ptr<AutochipMovePrimitiveMsg> createAutochipMovePrimitiveMsg(
     auto autochip_move_primitive_msg = std::make_unique<AutochipMovePrimitiveMsg>();
 
     auto position_params_msg = std::make_unique<MovePositionParams>();
-    auto dest_msg            = std::make_unique<PointMsg>();
-    dest_msg->set_x_meters(static_cast<float>(dest.x()));
-    dest_msg->set_y_meters(static_cast<float>(dest.y()));
+    auto dest_msg            = createPointMsg(Point(dest.x(), dest.y()));
     *(position_params_msg->mutable_destination()) = *dest_msg;
     position_params_msg->set_final_speed_meters_per_second(final_speed_meters_per_second);
     position_params_msg->set_slow(slow);
     *(autochip_move_primitive_msg->mutable_position_params()) = *position_params_msg;
 
-    auto final_angle_msg = std::make_unique<AngleMsg>();
-    final_angle_msg->set_radians(static_cast<float>(final_angle.toRadians()));
+    auto final_angle_msg                                  = createAngleMsg(final_angle);
     *(autochip_move_primitive_msg->mutable_final_angle()) = *final_angle_msg;
 
     autochip_move_primitive_msg->set_dribbler_speed_rpm(dribbler_speed_rpm);
@@ -123,16 +109,13 @@ std::unique_ptr<AutokickMovePrimitiveMsg> createAutokickMovePrimitiveMsg(
     auto autokick_move_primitive_msg = std::make_unique<AutokickMovePrimitiveMsg>();
 
     auto position_params_msg = std::make_unique<MovePositionParams>();
-    auto dest_msg            = std::make_unique<PointMsg>();
-    dest_msg->set_x_meters(static_cast<float>(dest.x()));
-    dest_msg->set_y_meters(static_cast<float>(dest.y()));
+    auto dest_msg            = createPointMsg(Point(dest.x(), dest.y()));
     *(position_params_msg->mutable_destination()) = *dest_msg;
     position_params_msg->set_final_speed_meters_per_second(final_speed_meters_per_second);
     position_params_msg->set_slow(slow);
     *(autokick_move_primitive_msg->mutable_position_params()) = *position_params_msg;
 
-    auto final_angle_msg = std::make_unique<AngleMsg>();
-    final_angle_msg->set_radians(static_cast<float>(final_angle.toRadians()));
+    auto final_angle_msg                                  = createAngleMsg(final_angle);
     *(autokick_move_primitive_msg->mutable_final_angle()) = *final_angle_msg;
 
     autokick_move_primitive_msg->set_dribbler_speed_rpm(dribbler_speed_rpm);
@@ -143,7 +126,7 @@ std::unique_ptr<AutokickMovePrimitiveMsg> createAutokickMovePrimitiveMsg(
     return autokick_move_primitive_msg;
 }
 
-std::unique_ptr<StopPrimitiveMsg> createStopPrimitiveMsg(StopTypeMsg stop_type)
+std::unique_ptr<StopPrimitiveMsg> createStopPrimitiveMsg(StopType stop_type)
 {
     auto stop_primitive_msg = std::make_unique<StopPrimitiveMsg>();
